backend/requirements.txt
PyJWT
passlib[bcrypt]
fastapi
uvicorn
sqlalchemy
boto3
python-multipart
